requirements.txt
astroid==2.4.2
boto3==1.14.51
botocore==1.17.63
click==7.1.2
colorama==0.4.4
decorator==4.4.2
dnspython==2.0.0
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.1
Pillow==8.0.0
psycopg2==2.8.6
pylint==2.6.0
python-dateutil==2.8.1
requests==2.9.1
s3transfer==0.3.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
validators==0.18.1
Werkzeug==1.0.1
wrapt==1.12.1
